recordcleaner/requirements.txt
asn1crypto==0.24.0
bcrypt==3.1.4
beautifulsoup4==4.6.0
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
cryptography==2.2.2
et-xmlfile==1.0.1
idna==2.6
inflect==0.3.1
jaconv==0.2.3
jdcal==1.4
numpy==1.14.3
openpyxl==2.5.3
pandas==0.23.0
paramiko==2.4.1
pip-autoremove==0.9.1
pyasn1==0.4.3
pycparser==2.18
PyNaCl==1.2.1
PyPDF2==1.26.0
python-dateutil==2.7.3
python-dotenv==0.8.2
pytz==2018.4
requests==2.19.1
six==1.11.0
sortedcontainers==2.0.2
tabulate==0.8.2
urllib3==1.22
Whoosh==2.7.4
xlrd==1.1.0
